Thread context through git remote ops and harden docker lifecycle
Propagate context.Context through git Clone/Fetch/Push and the
RemoteManager interface, detecting cancellation and deadlines first so
network errors preserve errors.Is while still redacting URL credentials.
Add ctx prechecks and cancellation-safe waits to the docker provider,
and split the embedded git manage/read files by concern.
